refactor: move the vit to the common utils dir, and make it be as general as possible, and
