+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
   # Ruff version.
-  rev: v0.12.1
+  rev: v0.12.2
   hooks:
     # Run the linter.
     - id: ruff
