--- src/jimm/models/common/__init__.py
@@ -1,3 +1,0 @@
-from .vit import VisionTransformerBase
-
-__all__ = ["VisionTransformerBase"]
