--- src/jimm/models/siglip.py
@@ -1,10 +1,0 @@
-from flax import nnx
-
-
-class SigLIPVisionTransformer(nnx.Module):
-    def __init__(
-        self,
-        input_resolution: int,
-        patch_size: int,
-    ):
-        pass
